+++ src/Webhook/Message/UpdateSubscriptionMessage.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Xsolla\SDK\Webhook\Message;
+
+class UpdateSubscriptionMessage extends CancelSubscriptionMessage
+{
+}